+++ fp-bindgen/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added end-to-end tests.
 - Fix #105: Correctly handle passing negative integers as primitives (outside
   MessagePack) between Rust and TypeScript.
+- Fix type of MessagePack-encoded 64-bit integers in TypeScript bindings.
 - Various smaller bugfixes.
 - **Deprecation**: `BindingsType::TsRuntime` is now deprecated in favor of
   `BindingsType::TsRuntimeWithExtendedConfig`.